+++ internal/writer/writer_test.go
@@ -1,7 +1,7 @@
 package writer
 
 /*
-   Copyright (c) 2025 SlideLang
+   Copyright (c) 2025 Misael Monterroca
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU Affero General Public License as published