+++ v2/examples/08_images/main.go
@@ -1,7 +1,7 @@
 /*
 MIT License
 
-Copyright (c) 2025 Misael Montero
+Copyright (c) 2025 Misael Monterroca
 
 Example: Image insertion with various positioning options